requirements.txt
streamlit>=1.32.0
pytest>=8.0.0
openai>=1.30.0
google-genai>=1.0.0
numpy>=1.26.0
